# api/openapi.yaml
...
openapi: 3.2.0
info:
  title: Layover Tower API
  version: 0.2.0
  summary: Control surface for a local-first agent factory.
  description: |
    The UI is purely a client of this API, so this document bounds what the UI can ever do.

    Everything is local-first and unauthenticated by default: the Tower binds to loopback. Do not
    expose it without putting something in front of it.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:7878
    description: The default local Tower.
...
paths:
...
  /runs:
    get:
      operationId: listRuns
      summary: Runs, live and historical.
      tags: [work]
      parameters:
        - name: status
          in: query
          required: false
          description: Return only runs in this state.
          schema:
            $ref: "#/components/schemas/RunStatus"
        - name: itinerary_id
          in: query
          required: false
          description: Return only runs belonging to this itinerary.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum number of runs to return, newest first.
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Matching runs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"
...
components:
  schemas:
...
    RunStatus:
      type: string
      description: |
        How a run ended, or that it has not. `stalled` is deliberately distinct from `failed`: a
        factory that quietly parks work forever is worse than one that crashes, so it has to be
        visible as its own outcome.
      enum: [running, succeeded, failed, stalled, cancelled]
    Run:
      type: object
      description: One supervised CLI execution.
      required: [run_id, itinerary_id, agent, status, started_at]
      properties:
        run_id:
          type: string
          description: Identifier of this run.
        itinerary_id:
          type: string
          description: The chain this run belongs to.
        agent:
          type: string
          description: Which agent was run.
        status:
          $ref: "#/components/schemas/RunStatus"
          description: How it ended, or that it has not.
        started_at:
          type: string
          format: date-time
          description: When the process was spawned.
        finished_at:
          type: [string, "null"]
          format: date-time
          description: When the process exited; null while it is still running.
        exit_code:
          type: [integer, "null"]
          format: int32
          description: Process exit code, when there was one.
        cost_usd:
          type: [number, "null"]
          format: double
          description: Null when the runner reported no cost, which the Tower logs loudly.
        hops_remaining:
          type: [integer, "null"]
          format: int32
          description: Hops left when this run was started, mirrored for display only.
    RunList:
      type: object
      description: Matching runs, newest first.
      required: [runs]
      properties:
        runs:
          type: array
          description: The runs.
          items:
            $ref: "#/components/schemas/Run"
...